requirements.txt
altgraph==0.17.5
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
future==1.0.0
iso8601==2.1.0
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.4.2
packaging==26.0
pandas==3.0.1
pefile==2024.8.26
pillow==12.1.1
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.2
pyparsing==3.3.2
pyserial==3.5
python-dateutil==2.9.0.post0
pywin32-ctypes==0.2.3
PyYAML==6.0.3
setuptools==82.0.0
six==1.17.0
tzdata==2025.3
